requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.8
django-axes==8.0.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
redis==5.2.1
